requirements.txt
flask
twilio
python-dotenv
requests
google-generativeai
